requirements.txt
gradio
transformers
pillow
python-docx
pytesseract
PyMuPDF
langdetect